Owned fishes displayed using Fish.kt
Titles, descriptions and the image of the fish is taken threw Fish.kt
