+++ app/src/main/java/com/github/k409/fitflow/ui/screens/inventory/InventoryScreen.kt
@@ -1,14 +1,11 @@
 package com.github.k409.fitflow.ui.screens.inventory
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Anchor
-import androidx.compose.material.icons.filled.Water
-import androidx.compose.material.icons.outlined.Anchor
-import androidx.compose.material.icons.outlined.Water
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
@@ -23,43 +20,14 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import com.github.k409.fitflow.R
 
-data class BottomNavigationItem(
-    val title: String,
-    val selectedIcon: ImageVector,
-    val unselectedIcon: ImageVector,
-)
-
 @Composable
 fun InventoryScreen(){
-    val ownedFishes = listOf(
-        Item("Clownfish",
-            "The fish from the movie \"Finding Nemo\"",
-            R.drawable.primary_fish),
-    )
-
-    val ownedDecorations = listOf(
-        Item("Plant",
-            "I don't know what type of plant this is",
-            R.drawable.plant),
-    )
-
-    val items = listOf(
-        BottomNavigationItem(
-            title = "Fishes",
-            selectedIcon = Icons.Filled.Water,
-            unselectedIcon = Icons.Outlined.Water,
-        ),
-        BottomNavigationItem(
-            title = "Decorations",
-            selectedIcon = Icons.Filled.Anchor,
-            unselectedIcon = Icons.Outlined.Anchor,
-        )
-    )
     var selectedItemIndex by rememberSaveable { mutableIntStateOf(0) }
+
     Surface(
         modifier = Modifier.fillMaxSize(),
         color = MaterialTheme.colorScheme.background,
@@ -89,6 +57,15 @@ fun InventoryScreen(){
                 }
             }
         ) {paddingValues ->
+            Box(
+                modifier = Modifier.fillMaxSize()
+            ){
+                Image(
+                    painter = painterResource(id = R.drawable.chest),
+                    contentDescription = "Background Image",
+                    modifier = Modifier.fillMaxSize(),
+                )
+            }
             Column(
                 modifier = Modifier
                     .fillMaxSize()
