+++ app/src/main/java/com/github/k409/fitflow/ui/screens/inventory/OwnedDecorations.kt
@@ -1,0 +1,9 @@
+package com.github.k409.fitflow.ui.screens.inventory
+
+import com.github.k409.fitflow.R
+
+val ownedDecorations = listOf(
+    Item("Plant",
+        "I don't know what type of plant this is",
+        R.drawable.plant),
+)
